add native host runner timeout
