+++ zkvm-host/bin/multi.rs
@@ -1,4 +1,4 @@
-use std::{fs, process::Command};
+use std::{fs, time::Duration};
 
 use anyhow::Result;
 use clap::Parser;
@@ -8,7 +8,7 @@ use host_utils::{
     get_proof_stdin, ProgramType,
 };
 use sp1_sdk::{utils, ExecutionReport, ProverClient};
-use zkvm_host::{convert_host_cli_to_args, precompile_hook, BnStats, ExecutionStats};
+use zkvm_host::{precompile_hook, run_native_host_runner, BnStats, ExecutionStats};
 
 pub const MULTI_BLOCK_ELF: &[u8] = include_bytes!("../../elf/validity-client-elf");
 
@@ -94,15 +94,7 @@ async fn main() -> Result<()> {
         fs::create_dir_all(&data_dir).unwrap();
 
         // Start the server and native client.
-        let metadata = cargo_metadata::MetadataCommand::new()
-            .exec()
-            .expect("Failed to get cargo metadata");
-        let target_dir = metadata.target_directory.join("release");
-        let args = convert_host_cli_to_args(&host_cli);
-        Command::new(target_dir.join("native_host_runner"))
-            .args(&args)
-            .spawn()?
-            .wait()?;
+        run_native_host_runner(&host_cli, Duration::from_secs(1)).await?;
     }
 
     // Get the stdin for the block.
